Use File::Spec to generate the template
We can't use the symlink name directly, because it could have
path separators in it. Break it apart with File::Spec::splitpath,
then concatenate it back together with catpath so we know the
temporary symlink will be created in the same directory as the
final name, and it will be a valid path.